Fix test dependency
